pyproject.toml
[tool.poetry]
name = "parse_uiks"
version = "0.1.0"
description = ""
authors = ["Artem Kanarev <aduchimergen@gmail.com>"]

[tool.poetry.dependencies]
python = "3.6"
django = "^2.0"
django-extensions = "^2.0"
pendulum = "^2.0"
requests = "^2.19"
beautifulsoup4 = "^4.6"
ipython = "^6.5"
django_q = "^0.9.4"
django-mptt = "^0.9.1"
arrow = "^0.12.1"
djangorestframework = "^3.8"
djangorestframework-gis = "^0.13.0"
djangorestframework-jsonp = "^1.0"
django-filter = "^2.0"

[tool.poetry.dev-dependencies]
